Working patch v2

+++ models/stock.go
@@ -1,8 +1,9 @@
 package models
 
+// Stock represents stock entity
 type Stock struct {
 	StockID int64  `json:"stockid"`
 	Name    string `json:"name"`
-	Price   int64  `json:"price"`
+	Price   int    `json:"price"`
 	Company string `json:"company"`
 }
